front end material and a few changes to the api
I commented out the database connection cause i couldn’t connect to it
i added a few custom return statements on the API side to generate fake
data. make sure that whatever comes from the database is returned in
this format.
On the front end, playlists are now clickable and so are song, but